agent_worker/requirements.txt
kafka-python
elasticsearch
groq
pdfminer.six
python-docx
pybreaker
python-dotenv
redis
jsonschema
pydantic-settings>=2.0.3  # Or latest version

